requirements.txt
packaging==21.0
PyYAML==5.4.1
zc.lockfile==2.0
schema==0.7.4
